fixing by comments
